requirements.txt
# Core Dependencies
psutil>=5.9.0
pywin32>=306
mss>=9.0.1
opencv-python>=4.8.0
matplotlib>=3.7.0
keyboard>=0.13.5
requests>=2.31.0
python-dotenv>=1.0.0

# API Integrations
openai>=1.0.0
anthropic>=0.7.0

# Audio Processing
pyaudio>=0.2.14
wave  # Built-in Python module

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# GUI (Optional - for legacy PyQt5 interface)
PyQt5>=5.15.10